requirements.txt
Pattern==2.6
nltk==3.0.5
pyenchant==1.6.6